+++ examples/next/src/pages/index.tsx
@@ -21,7 +21,7 @@ export default function Page() {
 
   return (
     <>
-      <h1>EVMts example</h1>
+      <h1>Evmts example</h1>
       <ConnectButton />
       {isConnected && (
         <>